requirements.txt
youtube-transcript-api
streamlit
google-genai
python-dotenv
pathlib
